Attempted to Add CoDrops + made own side navpanels
Ben - in map.html are the new nav panels. They operate via CSS
transition in custom.css and are triggered by calling “toggleLeftNav()”
and “toggleRightNav()”, as seen on the buttons in the nav on that page.
Please let me know if you need anything done on them at the moment.